requirements.txt
pymongo==3.5.1
python_dateutil==2.6.1
motor==1.1
tornado==4.5.2
requests=2.18.4
pycurl=7.43.0